fix(core): now_ms() — cast as_millis() to u64 (build was red)

+++ crates/core/src/session.rs
@@ -82,7 +82,7 @@ pub fn new_session_id() -> String {
 pub fn now_ms() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_millis())
+        .map(|d| d.as_millis() as u64)
         .unwrap_or(0)
 }
